# backend/src/main/java/com/dietiestate25backend/controller/OffertaController.java
package com.dietiestate25backend.controller;

import com.dietiestate25backend.dto.requests.AggiornaOffertaRequest;
import com.dietiestate25backend.dto.requests.CreaOffertaRequest;
import com.dietiestate25backend.model.Offerta;
import com.dietiestate25backend.service.OffertaService;
import com.dietiestate25backend.utils.TokenUtils;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/offerta")
@Validated
public class OffertaController {
    private final OffertaService offertaService;

    public OffertaController(OffertaService offertaService) {
        this.offertaService = offertaService;
    }

    @PostMapping("/aggiungi")
    public ResponseEntity<Void> aggiungiOfferta(@Valid @RequestBody CreaOffertaRequest request) {
        String uidCliente = TokenUtils.getUserSub();
        offertaService.aggiungiOfferta(request, uidCliente);
        return ResponseEntity.status(201).build();
    }

    @PatchMapping("/aggiorna")
    public ResponseEntity<Void> aggiornaStatoOfferta(@Valid @RequestBody AggiornaOffertaRequest request) {
        String uidCliente = TokenUtils.getUserSub();
        offertaService.aggiornaStatoOfferta(request, uidCliente);
        return ResponseEntity.ok().build();
    }

    @GetMapping("/riepilogoCliente")
    public ResponseEntity<List<Offerta>> riepilogoOfferteCliente() {
        String uidCliente = TokenUtils.getUserSub();
        List<Offerta> offerte = offertaService.riepilogoOfferteCliente(uidCliente);

        return ResponseEntity.ok(offerte);
    }

    @GetMapping("/riepilogoUtenteAgenzia")
    public ResponseEntity<List<Offerta>> riepilogoOfferteUtenteAgenzia() {
        TokenUtils.checkIfUtenteAgenzia();
        String idAgente = TokenUtils.getUserSub();
        List<Offerta> offerte = offertaService.riepilogoOfferteUtenteAgenzia(idAgente);

        return ResponseEntity.ok(offerte);
    }
}